fix bug on setState
Trying top fix this bug
Can't call setState on a component that is not yet mounted. This is a
no-op, but it might indicate a bug in your application. Instead, assign
to `this.state` directly or define a `state = {};` class property with
the desired state in the Calculator component.
What happens is that the component is not rendered when the state is
changend, or we can't change the state while the component won't be
rendered.